chore(eui): extract button constants for eui-vanilla
